chore: split api migration tests
